+++ src/safetywrap/_impl.py
@@ -176,6 +176,18 @@ class Ok(Result[T, E]):
         """
         return self._value
 
+    def raise_if_err(
+        self, msg: str, exc_cls: t.Type[Exception] = RuntimeError
+    ) -> T:
+        """Return `Ok` value or raise an error with the specified message.
+
+        The raised exception class may be specified with the `exc_cls`
+        keyword argument.
+
+        Alias for `Ok.expect`.
+        """
+        return self.expect(msg, exc_cls=exc_cls)
+
     def expect_err(
         self, msg: str, exc_cls: t.Type[Exception] = RuntimeError
     ) -> E:
@@ -305,6 +317,18 @@ class Err(Result[T, E]):
         """
         raise exc_cls(f"{msg}: {self._value}")
 
+    def raise_if_err(
+        self, msg: str, exc_cls: t.Type[Exception] = RuntimeError
+    ) -> T:
+        """Return `Ok` value or raise an error with the specified message.
+
+        The raised exception class may be specified with the `exc_cls`
+        keyword argument.
+
+        Alias for `Err.expect`.
+        """
+        return self.expect(msg, exc_cls=exc_cls)
+
     def expect_err(
         self, msg: str, exc_cls: t.Type[Exception] = RuntimeError
     ) -> E:
@@ -427,6 +451,18 @@ class Some(Option[T]):
         argument.
         """
         return self._value
+
+    def raise_if_err(
+        self, msg: str, exc_cls: t.Type[Exception] = RuntimeError
+    ) -> T:
+        """Unwrap and yield a `Some`, or throw an exception if `Nothing`.
+
+        The exception class may be specified with the `exc_cls` keyword
+        argument.
+
+        Alias of `Some.expect`.
+        """
+        return self.expect(msg, exc_cls=exc_cls)
 
     def filter(self, predicate: t.Callable[[T], bool]) -> Option[T]:
         """Return `Nothing`, or an option determined by the predicate.
@@ -577,6 +613,18 @@ class Nothing(Option[T]):
         argument.
         """
         raise exc_cls(msg)
+
+    def raise_if_err(
+        self, msg: str, exc_cls: t.Type[Exception] = RuntimeError
+    ) -> T:
+        """Unwrap and yield a `Some`, or throw an exception if `Nothing`.
+
+        The exception class may be specified with the `exc_cls` keyword
+        argument.
+
+        Alias of `Nothing.expect`.
+        """
+        return self.expect(msg, exc_cls=exc_cls)
 
     def filter(self, predicate: t.Callable[[T], bool]) -> Option[T]:
         """Return `Nothing`, or an option determined by the predicate.