+++ src/safetywrap/_interface.py
@@ -121,6 +121,21 @@ class _Result(t.Generic[T, E]):
         """
         raise NotImplementedError
 
+    def raise_if_err(
+        self, msg: str, exc_cls: t.Type[Exception] = RuntimeError
+    ) -> T:
+        """Return `Ok` value or raise an error with the specified message.
+
+        The raised exception class may be specified with the `exc_cls`
+        keyword argument.
+
+        The underlying error will be stringified and appended to the
+        provided message.
+
+        Alias of `expect`.
+        """
+        raise NotImplementedError
+
     def expect_err(
         self, msg: str, exc_cls: t.Type[Exception] = RuntimeError
     ) -> E:
@@ -261,6 +276,18 @@ class _Option(t.Generic[T]):
 
         The exception class may be specified with the `exc_cls` keyword
         argument.
+        """
+        raise NotImplementedError
+
+    def raise_if_err(
+        self, msg: str, exc_cls: t.Type[Exception] = RuntimeError
+    ) -> T:
+        """Unwrap and yield a `Some`, or throw an exception if `Nothing`.
+
+        The exception class may be specified with the `exc_cls` keyword
+        argument.
+
+        Alias of `expect`.
         """
         raise NotImplementedError